Change autofocus="true" to autofocus="" per HTML5 spec
http://www.w3.org/html/wg/drafts/html/master/infrastructure.html#boolean
-attribute

The presence of a boolean attribute on an element represents the true
value, and the absence of the attribute represents the false value.

If the attribute is present, its value must either be the empty string
or a value that is an ASCII case-insensitive match for the attribute's
canonical name, with no leading or trailing whitespace.