AutoForm.Utility.dateToDateStringUTC and AutoForm.Utility.dateToDateString now correctly handle dates with years < 1000. This fixes #774.
I added a TinyTest for this, which was a little tricky because we cannot
control the timezone the tests run in (the behavior of AutoForm.Utility.dateToDateString
is influenced by the timezone). I opted to have the test check what
timezone we're in, and assert appropriately (the 3 interesting cases
are: we're ahead of UTC, we're behind UTC, and we're in UTC). To test
these, I ran firefox with the TZ environment variable set:

TZ=UTC firefox
>> new Date().getTimezoneOffset()
0

TZ=Europe/Amsterdam firefox
>> new Date().getTimezoneOffset()
-60

TZ=America/Los_Angeles firefox
>> new Date().getTimezoneOffset()
420

The tests passed under all 3 instances of firefox.