Fixes multiple forms with same field name and value only saving first time.
This corrects an issue which came into play with the fix for #647.  I
encountered it when using forms in a forEach loop when each form has the
same field names and `autosave=true`.  When updating the same field
with the same value on multiple forms the value was not being submitted.

The same simple logic is being used as before, but now it uses an object
where the last value is looked up by formId (and not just globally).

This closes #1198